Refactor stations to simple component instead of page
